// src/shared/api/generated/model/teamProfilesResponseDtoMembersItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Yoclick API
 * API multi-centro de reservas en marca blanca. Contrato canónico para la app.
 * OpenAPI spec version: 0.1.0
 */
import type { TeamProfilesResponseDtoMembersItemVideo } from './teamProfilesResponseDtoMembersItemVideo.ts';

export type TeamProfilesResponseDtoMembersItem = {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  membershipId: string;
  isMe: boolean;
  fullName: string;
  /** @nullable */
  staffTitle: string | null;
  /** @nullable */
  video: TeamProfilesResponseDtoMembersItemVideo;
};